Rework tests based on changes to connect and Provider
